requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
Faker==29.0.0
exceptiongroup==1.2.2
executing==2.1.0
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pre-commit==3.8.0
Pygments==2.18.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
psycopg2-binary>=2.9
django-seed==0.3.1
